added action and dispatch for retrieving state from jwt and modified reducer to handle this action
