+++ src/actions/actionTypes.js
@@ -1,2 +1,5 @@
 // Login
 export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
+
+// Persistant State
+export const NO_USER_STATE = 'NO_USER_STATE';
