+++ src/actions/stateLossActions.js
@@ -1,0 +1,13 @@
+import * as types from './actionTypes';
+
+const stateLossActions = {
+
+  noUserState(token) {
+    return {
+      type: types.NO_USER_STATE,
+      userState: token,
+    };
+  },
+};
+
+module.exports = stateLossActions;
